Update gbn.h to allow header attribute calculation; Update readme.

+++ gbn.h
@@ -39,8 +39,8 @@ extern int errno;
 /*----- Go-Back-n packet format -----*/
 typedef struct {
 	uint8_t  type;            /* packet type (e.g. SYN, DATA, ACK, FIN)     */
-	uint8_t  seqnum;          /* sequence number of the packet              */
-	uint8_t  acknum;
+	uint32_t  seqnum;          /* sequence number of the packet              */
+	uint32_t  acknum;
     uint16_t checksum;        /* header and payload checksum                */
 	uint16_t  body_len;
     uint8_t data[DATALEN];    /* pointer to the payload                     */
